requirements.txt
splitwise~=2.3.0
xlsxwriter~=3.0.1
python_dateutil~=2.8.2
prompt_toolkit~=3.0.20
python-bidi~=0.4.2